+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed: don't attempt to derive bounding box for elements (including root `<svg>`)
+  where size / position attributes have units or %age values - e.g. `x="5%" width="10cm"`
+  etc. Previously this prevented processing many SVG examples using units or percentages.
+
 ## [0.9.2 - 2024-05-28]
 
 - Fixed (regression in 0.9.0): overriding compound attribute derived positions